DCA-4. CardList

+++ app/src/main/java/com/startandroid/biimka/discount_card_aggregator/Card.java
@@ -53,7 +53,7 @@ public class Card {
         return contentBarcode;
     }
 
-    public void setContentBarcode(int contentBarcode) {
+    public void setContentBarcode(long contentBarcode) {
         this.contentBarcode = contentBarcode;
     }
